#40: File store manager with common code for file-based storages like S3
